requirements.txt
# Core
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
pydantic==2.7.1

# Database
supabase==2.4.3
psycopg2-binary==2.9.9

# AI Models
openai==1.30.1
anthropic==0.26.0
google-generativeai==0.5.4
groq==0.8.0

# WhatsApp & Telegram
httpx==0.27.0
python-telegram-bot==21.3

# Google Sheets
gspread==6.1.2
google-auth==2.29.0

# Scheduling (auto sync)
apscheduler==3.10.4

# Testing
pytest==8.2.0
pytest-asyncio==0.23.7

# Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
